cloud_backend/requirements.txt
Flask==3.0.2
google-genai>=1.0.0
gunicorn==21.2.0
